Add new error types and handle some nonconformant files.
New Error Types: Updated errors for certain situations where the Invalid file error was being throw, this clarifies the cause of the error.

Handle some less conformant files: Many files in the wild features group blocks which are not fully parsed together, when this occurs they will incorectly be interpreted as empty. The change fixes this behaviour when the group block is not parsed together.