feat: handle files with nonconformant group block sizes
In some files found in the wild GroupStart blocks have a size which only contains the name of the group, this causes colors in that group to not be parsed. This change fixes that special case.

In addition, the change fixes an issue where some files feature a specified block size of 0 for GroupEnd blocks.